# improve: speed up GoPro processing