--- a/mapillary_tools/video_data_extraction/extractors/gopro_parser.py
+++ b/mapillary_tools/video_data_extraction/extractors/gopro_parser.py
@@ -1,3 +1,4 @@
+from __future__ import annotations
 import typing as T
 
 from ... import geo