Dataset:
```
❯ du -sh ~/Downloads/MyGoProVideos                              
 55G	~/Downloads/MyGoProVideos
```

Before:

```
❯ uv run -m mapillary_tools.commands --verbose process ~/Downloads/MyGoProVideos --desc_path /tmp/foo.json --skip_process_errors                                   
2025-03-19 15:17:56,729 - DEBUG   - CLI param: duplicate_angle: 5.0
2025-03-19 15:18:30,317 - INFO    - Found 27 videos and 1 errors after video limit checks
2025-03-19 15:18:30,318 - DEBUG   - Validating 51 metadatas
2025-03-19 15:18:31,449 - INFO    - Check the description file for details: /tmp/foo.json
2025-03-19 15:18:31,453 - INFO    -       51 gopro(s) read in total
2025-03-19 15:18:31,453 - INFO    -            27 gopro(s) (41813.7 MB) are ready to be uploaded
2025-03-19 15:18:31,453 - WARNING -            22 gopro(s) skipped due to MapillaryGPSNoiseError
2025-03-19 15:18:31,453 - WARNING -             1 gopro(s) skipped due to MapillaryStationaryVideoError
2025-03-19 15:18:31,453 - WARNING -             1 gopro(s) skipped due to MapillaryCaptureSpeedTooFastError
uv run -m mapillary_tools.commands --verbose process ~/Downloads/MyGoProVideo  146.53s user 12.71s system 453% cpu 35.135 total

```

After:
```
❯ time uv run -m mapillary_tools.commands --verbose process ~/Downloads/MyGoProVideos --desc_path /tmp/foo.json --skip_process_errors
2025-03-19 15:19:18,105 - DEBUG   - CLI param: duplicate_angle: 5.0
2025-03-19 15:19:24,250 - INFO    - Found 27 videos and 1 errors after video limit checks
2025-03-19 15:19:24,251 - DEBUG   - Validating 51 metadatas
2025-03-19 15:19:25,232 - INFO    - Check the description file for details: /tmp/foo.json
2025-03-19 15:19:25,236 - INFO    -       51 gopro(s) read in total
2025-03-19 15:19:25,236 - INFO    -            27 gopro(s) (41813.7 MB) are ready to be uploaded
2025-03-19 15:19:25,236 - WARNING -            22 gopro(s) skipped due to MapillaryGPSNoiseError
2025-03-19 15:19:25,236 - WARNING -             1 gopro(s) skipped due to MapillaryStationaryVideoError
2025-03-19 15:19:25,236 - WARNING -             1 gopro(s) skipped due to MapillaryCaptureSpeedTooFastError
uv run -m mapillary_tools.commands --verbose process ~/Downloads/MyGoProVideo  34.44s user 1.23s system 475% cpu 7.504 total
```